+++ crates/kira/src/manager/backend/resources/mixer.rs
@@ -7,7 +7,7 @@ use crate::{
 	clock::ClockTime,
 	dsp::Frame,
 	manager::{backend::context::Context, command::MixerCommand},
-	track::{effect::Effect, SubTrackId, Track, TrackBuilder, TrackId},
+	track::{SubTrackId, Track, TrackBuilder, TrackId},
 	tween::Tweener,
 };
 
@@ -24,17 +24,10 @@ impl Mixer {
 		sub_track_capacity: usize,
 		unused_sub_track_producer: Producer<Track>,
 		context: &Arc<Context>,
-		main_track_effects: Vec<Box<dyn Effect>>,
+		main_track_builder: TrackBuilder,
 	) -> Self {
 		Self {
-			main_track: Track::new(
-				{
-					let mut settings = TrackBuilder::new();
-					settings.effects = main_track_effects;
-					settings
-				},
-				context,
-			),
+			main_track: Track::new(main_track_builder, context),
 			sub_tracks: Arena::new(sub_track_capacity),
 			sub_track_ids: Vec::with_capacity(sub_track_capacity),
 			dummy_routes: vec![],